+++ src/compas_libigl/mapping.py
@@ -26,7 +26,7 @@ from compas_libigl._types_std import VectorInt  # noqa: F401
 from compas_libigl._types_std import VectorVectorInt  # noqa: F401
 
 
-def map_mesh(target_mesh, pattern_mesh, clip_boundaries=True, tolerance=1e-6):
+def map_mesh(target_mesh, pattern_mesh, clip_boundaries=True, simplify_borders=True, tolerance=1e-6):
     """
     Map a 2D pattern mesh onto a 3D target.
 
@@ -38,6 +38,8 @@ def map_mesh(target_mesh, pattern_mesh, clip_boundaries=True, tolerance=1e-6):
         A tuple of (vertices, faces) representing the pattern mesh.
     clip_boundaries : bool
         Whether to clip the pattern mesh to the boundaries of the target mesh.
+    simplify_borders : bool
+        Whether to simplify the border of the pattern mesh.
     tolerance : float
         The tolerance for point comparison, to remove duplicates.
 
@@ -63,14 +65,14 @@ def map_mesh(target_mesh, pattern_mesh, clip_boundaries=True, tolerance=1e-6):
 
     # Perform the mapping
     pv_numpy_copy, pf_numpy_cleaned, p_normals, pattern_is_boundary, pattern_groups = _mapping.map_mesh_with_automatic_parameterization(
-        v_numpy, f_numpy, pattern_v_numpy, pf, clip_boundaries, tolerance
+        v_numpy, f_numpy, pattern_v_numpy, pf, clip_boundaries, simplify_borders, tolerance
     )
 
     # Return the result as a tuple
     return pv_numpy_copy, pf_numpy_cleaned, p_normals, pattern_is_boundary, pattern_groups
 
 
-def map_pattern_to_mesh(name, mesh, clip_boundaries=True, tolerance=1e-6, pattern_u=16, pattern_v=16):
+def map_pattern_to_mesh(name, mesh, clip_boundaries=True, tolerance=1e-6, pattern_u=16, pattern_v=16, simplify_borders=True):
     """
     Map a 2D pattern mesh onto a 3D target.
 
@@ -108,6 +110,8 @@ def map_pattern_to_mesh(name, mesh, clip_boundaries=True, tolerance=1e-6, patter
         The number of pattern vertices in the u direction.
     pattern_v : int
         The number of pattern vertices in the v direction.
+    simplify_borders : bool
+        Whether to simplify the border of the pattern mesh.
 
     Returns
     -------
@@ -165,6 +169,6 @@ def map_pattern_to_mesh(name, mesh, clip_boundaries=True, tolerance=1e-6, patter
     pf = tessagon_mesh["face_list"]
 
     v, f = mesh.to_vertices_and_faces()
-    mapped_vertices, mapped_faces, mapped_normals, mapped_is_boundary, mapped_groups = map_mesh((v, f), (pv, pf), clip_boundaries=clip_boundaries, tolerance=tolerance)
+    mapped_vertices, mapped_faces, mapped_normals, mapped_is_boundary, mapped_groups = map_mesh((v, f), (pv, pf), clip_boundaries=clip_boundaries, simplify_borders=simplify_borders, tolerance=tolerance)
 
     return Mesh.from_vertices_and_faces(mapped_vertices, mapped_faces)